+++ src/backend/mouse_cache.rs
@@ -9,7 +9,7 @@ use crate::event::MouseButtonState;
 pub struct MouseCache {
 	mouse_buttons: BTreeMap<DeviceId, MouseButtonState>,
 	mouse_position: BTreeMap<(WindowId, DeviceId), PhysicalPosition<f64>>,
-	mouse_previous_position: BTreeMap<(WindowId, DeviceId), PhysicalPosition<f64>>,
+	mouse_prev_position: BTreeMap<(WindowId, DeviceId), PhysicalPosition<f64>>,
 }
 
 impl MouseCache {
@@ -17,8 +17,8 @@ impl MouseCache {
 		self.mouse_position.get(&(window_id, device_id)).copied()
 	}
 
-	pub fn get_previous_position(&self, window_id: WindowId, device_id: DeviceId) -> Option<PhysicalPosition<f64>> {
-		self.mouse_previous_position.get(&(window_id, device_id)).copied()
+	pub fn get_prev_position(&self, window_id: WindowId, device_id: DeviceId) -> Option<PhysicalPosition<f64>> {
+		self.mouse_prev_position.get(&(window_id, device_id)).copied()
 	}
 
 	pub fn get_buttons(&self, device_id: DeviceId) -> Option<&MouseButtonState> {
@@ -41,8 +41,8 @@ impl MouseCache {
 			},
 			WindowEvent::CursorMoved { device_id, position, .. } => {
 				let cached_position = self.mouse_position.entry((window_id, *device_id)).or_insert_with(|| [0.0, 0.0].into());
-				let cached_previous_position = self.mouse_previous_position.entry((window_id, *device_id)).or_insert_with(|| [0.0, 0.0].into());
-				*cached_previous_position = *cached_position;
+				let cached_prev_position = self.mouse_prev_position.entry((window_id, *device_id)).or_insert_with(|| [0.0, 0.0].into());
+				*cached_prev_position = *cached_position;
 				*cached_position = *position;
 			},
 			_ => {},
@@ -61,9 +61,9 @@ impl MouseCache {
 		for key in &keys {
 			self.mouse_position.remove(key);
 		}
-		let keys: Vec<_> = self.mouse_previous_position.keys().filter(|(_, x)| *x == device_id).copied().collect();
+		let keys: Vec<_> = self.mouse_prev_position.keys().filter(|(_, x)| *x == device_id).copied().collect();
 		for key in &keys {
-			self.mouse_previous_position.remove(key);
+			self.mouse_prev_position.remove(key);
 		}
 	}
 }